tensorflow now makes use of general frontend
